エラー発生時に読み込みが発生しないように修正した
